First working example with a queue
In this commit, we create an SQS queue via serverless.yml file, and 2 more functions.
One function reads list of hosts from the same json list used for scheduled scans, and puts them in the queue.
The second function simply prints out the next message in queue (event). That's it for now.

Note that this is a prototype.