Deprecate vautomator-serverless
vautomator-serverless is no longer supported. The Python package that the tool used to interact with the Tenable.io
API, tenable-io ( https://github.com/tenable/Tenable.io-SDK-for-Python ), was deprecated in 2020. In order
to fork this project and get it working, you'd need to update vautomator-serverless to use a different
Python package for interacting with Tenable.io